Render the LearnNavLinkCard outside
